+++ .gitignore
@@ -163,3 +163,6 @@ cython_debug/
 *vtp
 *vtk
 *vts
+
+# vscode
+.vscode
