+++ pnm_mctools/__init__.py
@@ -1,3 +1,2 @@
 from .ToolSet import MulticomponentTools                                 # noqa: F401
 from .NumericalDifferentiation import conduct_numerical_differentiation  # noqa: F401
-from .Reactions import LinearReaction                                    # noqa: F401
