+++ pnm_mctools/models/const_spheres_and_cylinders.py
