+++ pnm_mctools/test/test_MassTransfer.py
@@ -5,9 +5,11 @@ sys.path.append(str(parent_dir))
 
 import openpnm as op                                       # noqa: E402
 import scipy, scipy.linalg, scipy.sparse                   # noqa: E401, E402
-import testing.const_spheres_and_cylinders as geo_model    # noqa: E402
+import models.const_spheres_and_cylinders as geo_model    # noqa: E402
 import numpy as np                                         # noqa: E402
 from ToolSet import MulticomponentTools                    # noqa: E402
+import Operators as ops                                    # noqa: E402
+import BoundaryConditions as bc                            # noqa: E402
 
 Nx = 10
 Ny = 100000
@@ -25,36 +27,36 @@ network.regenerate_models()
 
 c = np.zeros((network.Np, 1))
 c_old = c.copy()
-bc = {}
-bc['left'] = {'prescribed': 1.}
-bc['right'] = {'prescribed': 0.}
 
 x = np.ndarray.flatten(c).reshape((c.size, 1))
 dx = np.zeros_like(x)
 x_old = x.copy()
 
-mt = MulticomponentTools(network=network, num_components=1, bc=bc)
-grad = mt.get_gradient_matrix()
-sum = mt.get_sum()
-ddt = mt.get_ddt(dt=0.0001)
+mt = MulticomponentTools(network=network, num_components=1)
+bc.set(mt, label='left', bc=1.)
+bc.set(mt, label='right', bc=0.)
+
+grad = ops.gradient(mt)
+sum = ops.sum(mt)
+ddt = ops.ddt(mt, dt=0.0001)
 D = np.ones((network.Nt, 1), dtype=float)
 
 J = ddt - sum(D, grad)
-J = mt.apply_bc(A=J)
+J = bc.apply(mt, A=J)
 
 for i in range(10):
     # timesteps
     x_old = x.copy()
 
     G = J * x - ddt * x_old
-    G = mt.apply_bc(x=x, b=G, type='Defect')
+    G = bc.apply(mt, x=x, b=G, type='Defect')
 
     for n in range(10):
         # iterations (should not take more than one!)
         dx[:] = scipy.sparse.linalg.spsolve(J, -G).reshape(dx.shape)
         x = x + dx
         G = J * x - ddt * x_old
-        G = mt.apply_bc(x=x, b=G, type='Defect')
+        G = bc.apply(mt, x=x, b=G, type='Defect')
         G_norm = np.linalg.norm(np.abs(G), ord=2)
         if G_norm < 1e-6:
             break
